Add spotlight prop to Tooltip, break getDotPosition out
