+++ packages/reactv2/src/components/Tooltip/index.tsx
@@ -8,18 +8,20 @@ import { Box } from '../Box'
 import { Button } from '../Button'
 import { Dot } from './Dot'
 import { Text } from '../Text'
+import { getDotPosition } from './getDotPosition'
 import { mapTooltipPropsToRadixProps } from './mapTooltipPropsToPopoverProps'
 
 interface MergedRadixPopoverProps
   extends Pick<Popover.PopoverProps, 'defaultOpen' | 'modal' | 'onOpenChange' | 'open'>,
     Omit<Popover.PopoverContentProps, 'align' | 'asChild'> {}
 export interface TooltipProps extends MergedRadixPopoverProps {
-  anchor?: string
-  style?: React.CSSProperties
   align?: Popover.PopoverContentProps['align'] | 'before' | 'after'
+  anchor?: string
+  spotlight?: boolean
+  style?: React.CSSProperties
 }
 
-export function Tooltip({ anchor, style, ...props }: TooltipProps) {
+export function Tooltip({ anchor, spotlight = false, style, ...props }: TooltipProps) {
   const { node: contentNode, rect: contentRect, ref: contentRef } = useBoundingClientRect()
   const [alignAttr, setAlignAttr] = useState(props.align)
   const [sideAttr, setSideAttr] = useState(props.side)
@@ -61,133 +63,96 @@ export function Tooltip({ anchor, style, ...props }: TooltipProps) {
 
   if (anchorElementRef == null) return null
 
-  function getDotPosition() {
-    const currentSide = sideAttr ?? 'bottom'
-    const dotProps = {}
+  const anchorRect = anchorElementRef.current.getBoundingClientRect()
 
-    // Radix's collision system isn't aware of our custom before|after align
-    const currentAlign = (() => {
-      if (['after', 'before'].includes(props.align)) {
-        if (alignAttr == 'start') {
-          return 'before'
-        } else if (alignAttr == 'end') {
-          return 'after'
-        }
-      }
-
-      return props.align
-    })()
-
-    const dotOffset = '-24px'
-
-    const oppositeSides = {
-      top: 'bottom',
-      right: 'left',
-      bottom: 'top',
-      left: 'right',
-    }
-
-    /* 
-      Rules:
-        - Dot is opposite to side prop (e.g. side=left -> dot=right)
-        - align=before|end -> Dot goes to highest extent (right/bottom) of align-axis
-        - align=after|start -> Dot goes to lowest extent (left/top) of align-axis
-        - align=center -> Dot goes to the center
-    */
-
-    dotProps[oppositeSides[currentSide]] = dotOffset
-
-    if (['before', 'end'].includes(currentAlign)) {
-      if (['top', 'bottom'].includes(currentSide)) {
-        dotProps['right'] = dotOffset
-      } else {
-        dotProps['bottom'] = dotOffset
-      }
-    } else if (['after', 'start'].includes(currentAlign)) {
-      if (['top', 'bottom'].includes(currentSide)) {
-        dotProps['left'] = dotOffset
-      } else {
-        dotProps['top'] = dotOffset
-      }
-    } else {
-      // The only option left is align=center
-      if (['top', 'bottom'].includes(currentSide)) {
-        dotProps['left'] = `calc(50% + ${dotOffset})`
-      } else {
-        dotProps['top'] = `calc(50% + ${dotOffset})`
-      }
-    }
-
-    return dotProps
+  // Needs to be a layoutEffect so it can check _after_ the anchor renders
+  let anchorRadius = '0'
+  if (typeof window !== 'undefined') {
+    anchorRadius = window.getComputedStyle(anchorElementRef.current).borderRadius
   }
 
-  const dotPosition = getDotPosition()
+  const dotPosition = getDotPosition({ props, alignAttr, sideAttr })
 
   return (
     <Popover.Root defaultOpen={true} {...rootProps}>
-      <Popover.Anchor virtualRef={anchorElementRef} />
+      <Popover.Anchor virtualRef={anchorElementRef} style={{ borderRadius: '10px' }} />
       <Popover.Portal>
-        <Popover.Content asChild {...contentProps} ref={contentRef}>
-          <Box
-            backgroundColor="white"
-            borderRadius="md"
-            padding={5}
-            style={{
-              boxShadow: '0px 4px 20px rgba(0, 0, 0, 0.1)',
-              display: 'flex',
-              flexDirection: 'column',
-              position: 'relative',
-              width: '300px',
-              ...style,
-            }}
-          >
-            <Dot style={dotPosition} />
-
-            {/* Image placeholder */}
+        <>
+          {spotlight && (
             <Box
-              backgroundColor="gray900"
-              borderRadius="md"
-              mb={5}
-              mt={-5}
-              mx={-5}
               style={{
-                aspectRatio: '2',
-                borderBottomLeftRadius: 0,
-                borderBottomRightRadius: 0,
+                borderRadius: anchorRadius,
+                boxShadow: '0 0 0 2000px rgb(0 0 0 / 0.5)',
+                height: anchorRect.height,
+                left: anchorRect.left,
+                position: 'absolute',
+                top: anchorRect.top,
+                width: anchorRect.width,
               }}
             />
-
-            <Text.Body1 fontWeight="bold" mb={1}>
-              {title}
-            </Text.Body1>
-            <Text.Body2>{subtitle}</Text.Body2>
-
+          )}
+          <Popover.Content asChild {...contentProps} ref={contentRef}>
             <Box
-              pt={4}
+              backgroundColor="white"
+              borderRadius="md"
+              padding={5}
               style={{
+                boxShadow: '0px 4px 20px rgba(0, 0, 0, 0.1)',
                 display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'space-between',
+                flexDirection: 'column',
+                position: 'relative',
+                width: '300px',
+                ...style,
               }}
             >
-              <Text.Body2 fontWeight="demibold">1/4</Text.Body2>
-              <Button.Primary title={primaryButtonTitle ?? 'Ok'} />
-            </Box>
+              <Dot style={dotPosition} />
 
-            <Popover.Close
-              aria-label="Close"
-              style={{
-                background: 'transparent',
-                border: 0,
-                position: 'absolute',
-                top: 8,
-                right: 0,
-              }}
-            >
-              <XMarkIcon height="20" fill="black" />
-            </Popover.Close>
-          </Box>
-        </Popover.Content>
+              {/* Image placeholder */}
+              <Box
+                backgroundColor="gray900"
+                borderRadius="md"
+                mb={5}
+                mt={-5}
+                mx={-5}
+                style={{
+                  aspectRatio: '2',
+                  borderBottomLeftRadius: 0,
+                  borderBottomRightRadius: 0,
+                }}
+              />
+
+              <Text.Body1 fontWeight="bold" mb={1}>
+                {title}
+              </Text.Body1>
+              <Text.Body2>{subtitle}</Text.Body2>
+
+              <Box
+                pt={4}
+                style={{
+                  display: 'flex',
+                  alignItems: 'center',
+                  justifyContent: 'space-between',
+                }}
+              >
+                <Text.Body2 fontWeight="demibold">1/4</Text.Body2>
+                <Button.Primary title={primaryButtonTitle ?? 'Ok'} />
+              </Box>
+
+              <Popover.Close
+                aria-label="Close"
+                style={{
+                  background: 'transparent',
+                  border: 0,
+                  position: 'absolute',
+                  top: 8,
+                  right: 0,
+                }}
+              >
+                <XMarkIcon height="20" fill="black" />
+              </Popover.Close>
+            </Box>
+          </Popover.Content>
+        </>
       </Popover.Portal>
     </Popover.Root>
   )
